requirements.txt
requests
kafka-python
pyspark
psycopg2-binary
